requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.2
fastapi-cli==0.0.5
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.121
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.1
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.1
psycopg2==2.9.9
pyarrow==17.0.0
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.0.1
yarl==1.11.1
